+++ src/somesy/pom_xml/__init__.py
@@ -1,0 +1,1 @@
+"""Somesy implementation for Java Maven pom.xml."""